Created tests for a method connection in JDBCDatabaseManager. Start create a command Create. Change all exception type from Exception to SQLException

+++ src/main/java/model/JDBCDatabaseManager.java
@@ -1,6 +1,7 @@
 package model;
 
 import java.sql.*;
+import java.util.ArrayList;
 import java.util.HashSet;
 
 public class JDBCDatabaseManager implements DatabaseManager {
@@ -11,71 +12,78 @@ public class JDBCDatabaseManager implements DatabaseManager {
     }
 
     @Override
-    public void connect(String databaseName, String username, String password) throws Exception {
+    public void connect(String databaseName, String userName, String password) throws SQLException {
         closeIfWasConnected();
 
         String url = "jdbc:postgresql://localhost:5432/" + databaseName;
         try {
-            connection = DriverManager.getConnection(url, username, password);
+            connection = DriverManager.getConnection(url, userName, password);
         } catch (SQLException e) {
-            if (e.getMessage().equals("FATAL: password authentication failed for user \"" + username + "\"")) {
-                throw new Exception("Username or password are incorrect.", e);
+            if (e.getMessage().equals("FATAL: password authentication failed for user \"" + userName + "\"")) {
+                throw new SQLException("Username or password are incorrect.", e);
             } else if (e.getMessage().equals("FATAL: database \"" + databaseName + "\" does not exist")) {
-                throw new Exception("Database " + databaseName + " does not exist.", e);
+                throw new SQLException("Database " + databaseName + " does not exist.", e);
             } else {
                 throw e;
             }
         }
     }
 
-    private void closeIfWasConnected() throws Exception {
+    private void closeIfWasConnected() throws SQLException {
         if (isConnected()) {
             try {
                 connection.close();
-            } catch (Exception e) {
-                throw new Exception("Previous connection was not close. Try again or reset SQLCmd.", e);
+            } catch (SQLException e) {
+                throw new SQLException("Previous connection was not close. Try again or reset SQLCmd.", e);
             }
         }
     }
 
     @Override
     public HashSet<String> tables() throws SQLException {
-        String sql = "SELECT table_name FROM information_schema.tables\n" +
-                "WHERE table_schema NOT IN ('information_schema', 'pg_catalog')\n" +
-                "AND table_schema IN('public');";
-        HashSet<String> tableNames = new HashSet<>();
+        if (isConnected()) {
+            String sql = "SELECT table_name FROM information_schema.tables\n" +
+                    "WHERE table_schema NOT IN ('information_schema', 'pg_catalog')\n" +
+                    "AND table_schema IN('public');";
+            HashSet<String> tableNames = new HashSet<>();
 
-        try (Statement statement = connection.createStatement();
-             ResultSet resultSet = statement.executeQuery(sql))
-        {
-            while (resultSet.next()) {
-                tableNames.add(resultSet.getString(1));
+            try (Statement statement = connection.createStatement();
+                 ResultSet resultSet = statement.executeQuery(sql)) {
+                while (resultSet.next()) {
+                    tableNames.add(resultSet.getString(1));
+                }
+                return tableNames;
             }
-            return tableNames;
+        } else {
+            throw new SQLException("You are not connected to database");
         }
     }
 
     @Override
-    public void clear(String tableName) throws Exception {
-        try (Statement statement = connection.createStatement()) {
-            String sql = "DELETE FROM " + tableName + ";";
-            statement.executeUpdate(sql);
-        } catch (SQLException e) {
-            if (e.getMessage().equals("ERROR: relation \"" + tableName + "\" does not exist\n" + "  Позиция: 13")) {
-                throw new Exception("Table " + tableName + " does not exist.", e);
-            } else {
-                throw e;
+    public void clear(String tableName) throws SQLException {
+        if (isConnected()) {
+            try (Statement statement = connection.createStatement()) {
+                String sql = "DELETE FROM " + tableName + ";";
+                statement.executeUpdate(sql);
+            } catch (SQLException e) {
+                if (e.getMessage().equals("ERROR: relation \"" + tableName + "\" does not exist\n" + "  Позиция: 13")) {
+                    throw new SQLException("Table " + tableName + " does not exist.", e);
+                } else {
+                    throw e;
+                }
             }
+        } else {
+            throw new SQLException("You are not connected to database");
         }
     }
 
     @Override
-    public void exit() throws Exception {
+    public void exit() throws SQLException {
         if (isConnected()) {
             try {
                 connection.close();
             } catch (SQLException e) {
-                throw new Exception("Connection was not close.", e);
+                throw new SQLException("Connection was not close.", e);
             }
         }
     }
@@ -101,7 +109,7 @@ public class JDBCDatabaseManager implements DatabaseManager {
     }
 
     @Override
-    public void create(String command) {
+    public void create(String tableName, ArrayList<String> namesAndTypesOfColumns) {
 
     }
 
@@ -111,24 +119,3 @@ public class JDBCDatabaseManager implements DatabaseManager {
     }
 }
 
-
-//        String url = "jdbc:postgresql://localhost:5432/sqlcmdDB?currentSchema=sqlcmdDB";
-//        String login = "postgres";
-//        String password = "sup3r42pass";
-//        Connection connection = null;
-//
-//        String sql = "CREATE TABLE public.COMPANY " +
-//                "(ID INT PRIMARY KEY     NOT NULL," +
-//                " NAME           TEXT    NOT NULL, " +
-//                " AGE            INT     NOT NULL, " +
-//                " ADDRESS        CHAR(50), " +
-//                " SALARY         REAL);";
-//
-//        try {
-//            connection = DriverManager.getConnection(url, login, password);
-//            Statement statement = connection.createStatement();
-//            statement.execute(sql);
-//
-//        } catch (SQLException e) {
-//            e.printStackTrace();
-//        }
